requirements.txt
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
pyyaml>=6.0
joblib>=1.3.0
scikit-learn>=1.3.0
pyarrow>=14.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0